Moved Classes to classes folder and Interface to interfaces folder
