+++ src/classes/PassesItem.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\classes;
 
 class PassesItem extends BaseItem
 {
