+++ src/interfaces/DayIncrementInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\interfaces;
 
 interface DayIncrementInterface
 {
